+++ src/badge-groupings/dto/load-badge-grouping.dto.ts
@@ -1,0 +1,8 @@
+export class LoadBadgeGroupingDto {
+  id: number;
+  name: string;
+  description?: string;
+  position: number;
+  system: boolean;
+  forumUUID: string;
+}
