+++ src/badge-types/dto/load-badge-type.dto.ts
@@ -1,0 +1,6 @@
+export class LoadBadgeTypeDto {
+  id: number;
+  name: string;
+  sortOrder: number;
+  forumUUID: string;
+}
